fix: naming inconsistencies
