+++ src/hooks/useStorage/index.ts
@@ -1,4 +1,4 @@
 import useStorage from './useStorage'
 
-export * from './useStorage'
+export * from './types'
 export default useStorage
