+++ src/hooks/useStorage/useStorage.ts
@@ -1,20 +1,6 @@
 import { useCallback, useState } from 'react'
 
-interface Storage {
-  getItem(key: string): string | null
-  setItem(key: string, value: string): void
-  removeItem(key: string): void
-}
-
-export type UseStorageProps = {
-  storage: Storage
-  key: string
-  defaultValue?: any
-  isStringStorage?: boolean
-  setInit?: (item: any) => any
-}
-
-export type UseStorageReturnType = [any, (item: any) => void, () => void]
+import { UseStorageProps, UseStorageReturnType } from './types'
 
 const setInitDefault = (item: any): any => item
